pyproject.toml
[project]
name = "chat"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "grpcio (==1.72.0)",
    "grpcio-tools (==1.72.0)",
    "azure-messaging-webpubsubservice (==1.2.2)",
    "websockets (>=15.0.1,<16.0.0)",
    "python-dotenv (==1.1.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "azure-communication-identity (==1.5.0)",
    "notebook (>=7.4.7,<8.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
